Move error to typeck::check
